+++ NEWS.md
@@ -1,3 +1,9 @@
+# sjstats 0.13.1
+
+## Bug fixes
+
+* `mode_frame()` did not work properly for variables that were standardized with `scale()`.
+
 # sjstats 0.13.0
 
 ## General